+++ package.json
@@ -39,6 +39,7 @@
 	"author": "Fran Mendez <fmvilas@gmail.com> (fmvilas.com)",
 	"license": "Apache-2.0",
 	"dependencies": {
+		"@asyncapi/parser": "^2.0.0-next-major.11",
 		"js-yaml": "^3.14.1"
 	},
 	"devDependencies": {
@@ -52,11 +53,11 @@
 		"conventional-changelog-conventionalcommits": "^4.6.3",
 		"cross-env": "^7.0.3",
 		"jest": "^27.5.1",
+		"markdown-toc": "^1.2.0",
 		"semantic-release": "^19.0.3",
 		"ts-jest": "^27.1.3",
 		"ts-node": "^10.7.0",
-		"typescript": "^4.6.2",
-		"markdown-toc": "^1.2.0"
+		"typescript": "^4.6.2"
 	},
 	"release": {
 		"branches": [
